EltwiseData: refactoring
